requirements-common.txt
aiozmq
async_timeout
bs4
django
django-allauth
django-dirtyfields
django-honeypot
django-mailer
#django-rest-framework-rules
django-simple-captcha
django-statici18n
django-treebeard
django_translated_fields
djangorestframework
docutils
lxml
Pillow
psycopg2-binary
python-dateutil
python-magic
rules
uwsgi
uwsgi-tasks
yarl
